backend_flask/app/requirements.txt
Flask
Flask-Login
mysql-connector-python
reportlab
web3==6.15.0
eth-tester
Flask-Mail
flask-cors
Flask-Talisman
